# api/openapi.yaml
openapi: 3.1.0
info:
  title: Momentra Core API (HDAM)
  version: 1.0.0
  description: |
    API-first Specification for **Momentra** (Historical Digital Asset Management - HDAM).
    Binds digital assets, links, notes, and events to historical event dates with uncertainty support,
    multi-tenant workspace isolation, and high-performance timeline queries.
  contact:
    name: Momentra Engineering Architecture Team
    email: api@momentra.app
  license:
    name: Proprietary
servers:
  - url: https://momentra.app/api/v1
    description: Production Environment
  - url: http://localhost:3000/api/v1
    description: Local Development Server
...
security:
  - cookieAuth: []
    workspaceHeader: []
paths:
...
  /timeline:
    get:
      summary: Query Historical Timeline with Density Clustering & Bounding Box
      description: |
        Returns aggregated time buckets and clustered items for the viewport.
        Uses GiST Index Range Overlapping (&&) for sub-second responses.
      operationId: getTimeline
      tags: [Timeline]
      parameters:
        - name: from
          in: query
          required: true
          description: Viewport minimum date (ISO 8601 UTC)
          schema:
            type: string
            format: date-time
            example: 1932-01-01T00:00:00Z
        - name: to
          in: query
          required: true
          description: Viewport maximum date (ISO 8601 UTC)
          schema:
            type: string
            format: date-time
            example: 1950-12-31T23:59:59Z
        - name: granularity
          in: query
          required: true
          description: Zoom scale of timeline
          schema:
            type: string
            enum: [decade, year, month, day]
            default: year
        - name: calendar
          in: query
          description: Preferred display calendar standard (BE = พ.ศ., CE = ค.ศ.)
          schema:
            type: string
            enum: [be, ce]
            default: be
        - name: tag_ids
          in: query
          description: Filter items by tag UUIDs
          schema:
            type: array
            items:
              type: string
              format: uuid
      responses:
        '200':
          description: Timeline view data with clusters and milestones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineResponse'

  # ==========================================================================
  # 7. Search Engine (Faceted & Full-Text)
  # ==========================================================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: momentra_session
      description: HTTP-only Secure session cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for Guest Share access or API integrations
    workspaceHeader:
      type: apiKey
      in: header
      name: X-Workspace-ID
      description: Target workspace UUID context for Multi-Tenancy
...
  schemas:
...
    ItemResponse:
      type: object
      required: [id, workspace_id, type, title, event_start, event_end, date_precision, is_circa]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [asset, link, note, event]
        title:
          type: string
          example: การเปลี่ยนแปลงการปกครอง 24 มิถุนายน 2475
        description:
          type: string
          nullable: true
        event_start:
          type: string
          format: date-time
          example: 1932-06-24T05:00:00Z
        event_end:
          type: string
          format: date-time
          example: 1932-06-24T18:00:00Z
        date_precision:
          type: string
          enum: [year, month, day, datetime]
          example: datetime
        is_circa:
          type: boolean
          example: false
        sort_key:
          type: integer
          format: int64
        display_date_be:
          type: string
          example: 24 มิถุนายน 2475 เวลา 05:00 น.
        display_date_ce:
          type: string
          example: 24 June 1932 05:00 UTC
        dublin_core:
          type: object
        asset:
          $ref: '#/components/schemas/AssetResponse'
        link:
          $ref: '#/components/schemas/LinkResponse'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
...
    AssetResponse:
      type: object
      required: [storage_key, mime_type, size_bytes, checksum_sha256, status]
      properties:
        storage_key:
          type: string
        original_filename:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        checksum_sha256:
          type: string
        status:
          type: string
          enum: [uploading, processing, ready, failed, quarantined]
...
    LinkResponse:
      type: object
      required: [url, domain, is_broken]
      properties:
        url:
          type: string
        domain:
          type: string
        og_title:
          type: string
          nullable: true
        og_description:
          type: string
          nullable: true
        og_image:
          type: string
          nullable: true
        is_broken:
          type: boolean

    # Timeline Response
    TimelineResponse:
      type: object
      required: [granularity, calendar, buckets, milestones]
      properties:
        granularity:
          type: string
          enum: [decade, year, month, day]
        calendar:
          type: string
          enum: [be, ce]
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/TimelineBucket'
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/MilestoneResponse'
    TimelineBucket:
      type: object
      required: [bucket_key, display_label, count, items]
      properties:
        bucket_key:
          type: string
          example: "1932"
        display_label:
          type: string
          example: พ.ศ. 2475 (1932)
        count:
          type: integer
          example: 4
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemResponse'
    MilestoneResponse:
      type: object
      required: [id, title, target_date_start, target_date_end, color]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: การอภิวัฒน์สยาม 2475
        target_date_start:
          type: string
          format: date-time
        target_date_end:
          type: string
          format: date-time
        color:
          type: string
          example: '#DC2626'
...
    TagResponse:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: การเมืองและประชาธิปไตย
        color:
          type: string
          example: '#EF4444'
...